预约、取消预约 查看预约记录

+++ src/main/java/cn/njxz/fitness/controller/RecordController.java
@@ -1,19 +1,30 @@
 package cn.njxz.fitness.controller;
 
 import cn.njxz.fitness.constants.Constants;
+import cn.njxz.fitness.model.Course;
 import cn.njxz.fitness.model.Record;
+import cn.njxz.fitness.model.RecordVO;
 import cn.njxz.fitness.model.User;
+import cn.njxz.fitness.service.CourseService;
 import cn.njxz.fitness.service.IRecordService;
+import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.PageInfo;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
 
 /**
  * @author yue.wu
@@ -25,12 +36,15 @@ import java.time.LocalDateTime;
 public class RecordController {
     @Autowired
     private IRecordService recordService;
+    @Autowired
+    private CourseService courseService;
 
     @PostMapping("/reserve")
     @ResponseBody
-    public String reserve(Integer cId, Boolean flag, HttpServletRequest request){
+    public HashMap reserve(Integer cId, Boolean flag, HttpServletRequest request){
         HttpSession session = request.getSession();
         User user= (User)session.getAttribute("user");
+        HashMap map = new HashMap(2);
         if(null!=user){
             int uid = user.getUId();
             if(cId!=null){
@@ -39,26 +53,59 @@ public class RecordController {
                 record.setrCId(cId);
                 record.setrUId(uid);
                 record.setrTime(LocalDateTime.now());
+                Record record1 = recordService.findRecordByUidAndCid(uid,cId);
                 if (flag==false){
-                    int insertResult =recordService.addRecord(record);
-                    if (insertResult > Constants.OPERATE_FAILURE) {
-                        return "{\"msg\":\"ok\"}";
+                    if(record1==null) {
+                        int insertResult = recordService.addRecord(record);
+                        if (insertResult > Constants.OPERATE_FAILURE) {
+                            map.put("msg", "ok");
+                            map.put("flag", true);
+                            return map;
+                        }
                     }
                 }else if(flag==true){
                     //取消预约
-                    Record record1 = recordService.findRecordByUidAndCid(uid,cId);
-                    int deleteResult = recordService.deleteRecord(record1.getrId());
-                    if (deleteResult > Constants.OPERATE_FAILURE) {
-                        return "{\"msg\":\"ok\"}";
+                    if(record1!=null) {
+                        int deleteResult = recordService.deleteRecord(record1.getrId());
+                        if (deleteResult > Constants.OPERATE_FAILURE) {
+                            map.put("msg", "ok");
+                            map.put("flag", false);
+                            return map;
+                        }
                     }
                 }
             }
         }
 
-        return "{\"msg\":\"no\"}";
+        return map;
     }
 
+    @GetMapping("/findAllRecord")
+    public String findAllRecord(HttpServletRequest request,@RequestParam(defaultValue="1")Integer pageNum) {
+        HttpSession session = request.getSession();
+        User user = (User) session.getAttribute("user");
+        PageHelper.startPage(pageNum, 8);
+        List<Record> records = recordService.findRecordByUid(user.getUId());
+        List<RecordVO> recordVOList = new ArrayList<>();
+        RecordVO recordVO = new RecordVO();
+        records.stream().map(e -> e.getrCId()).forEach(e -> {
+            Course course = courseService.findCourseById(e);
+            BeanUtils.copyProperties(course,recordVO);
+            recordVO.setRId(e);
+            recordVOList.add(recordVO);
+        });
+        if (records != null) {
+            PageInfo<RecordVO> pageInfo = new PageInfo<>(recordVOList);
+            request.setAttribute("pageInfo", pageInfo);
+        }
+        return "templates/record";
+    }
 
+    @GetMapping("/cancelCourse")
+    public String cancelCourse(@RequestParam("rId")int rId) {
+
+        return "redirect:/record/findAllRecord";
+    }
 
 
 }
